+++ README.md
@@ -1848,7 +1848,7 @@ See [Language Definition](https://expr-lang.org/docs/language-definition).
 - `select` ... Select from candidates. `func(message string, candidates []string, default string) string`
 - `basename` ... [filepath.Base](https://pkg.go.dev/path/filepath#Base)
 - `time` ... Converts the given string or number to `time.Time{}`.
-- `faker.*` ... Generate fake data using [Faker](https://pkg.go.dev/github.com/k1LoW/runn/builtin#Faker) ).
+- `faker.*` ... Generate fake data using [Faker](https://pkg.go.dev/github.com/k1LoW/runn/internal/builtin#Faker) ).
 
 ## Option
