+++ internal/builtin/compare.go
