+++ internal/builtin/compare_test.go
