+++ internal/builtin/diff.go
