+++ internal/builtin/diff_test.go
