+++ internal/builtin/intersect_test.go
