+++ internal/expr/expr.go
@@ -14,7 +14,7 @@ import (
 	"github.com/expr-lang/expr/parser/lexer"
 	"github.com/goccy/go-yaml"
 	"github.com/k1LoW/expand"
-	"github.com/k1LoW/runn/builtin"
+	"github.com/k1LoW/runn/internal/builtin"
 	"github.com/k1LoW/runn/internal/deprecation"
 	"github.com/k1LoW/runn/internal/exprtrace"
 	"github.com/xlab/treeprint"
