+++ option.go
@@ -18,7 +18,7 @@ import (
 	"github.com/elk-language/go-prompt"
 	pstrings "github.com/elk-language/go-prompt/strings"
 	"github.com/k1LoW/duration"
-	"github.com/k1LoW/runn/builtin"
+	"github.com/k1LoW/runn/internal/builtin"
 	"github.com/k1LoW/runn/internal/expr"
 	"github.com/k1LoW/runn/internal/store"
 	"github.com/k1LoW/sshc/v4"
